Move stuff to serialize.rs
