+++ module-system/module-implementations/sov-evm/src/evm/mod.rs
@@ -1,5 +1,3 @@
-use std::io::Read;
-
 use borsh::{BorshDeserialize, BorshSerialize};
 use derive_more::{From, Into};
 use revm::primitives::specification::SpecId;
@@ -11,6 +9,7 @@ pub(crate) mod db;
 mod db_commit;
 pub(crate) mod db_init;
 pub(crate) mod executor;
+mod serialize;
 #[cfg(test)]
 pub(crate) mod test_helpers;
 #[cfg(test)]
@@ -77,27 +76,6 @@ pub(crate) fn contract_address(result: ExecutionResult) -> Option<B160> {
 /// EVM SpecId and their activation block
 #[derive(Debug, PartialEq, Clone, Copy, From, Into)]
 pub struct SpecIdWrapper(SpecId);
-
-impl BorshSerialize for SpecIdWrapper {
-    fn serialize<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()> {
-        let value = self.0 as u8;
-        value.serialize(writer)
-    }
-}
-
-impl BorshDeserialize for SpecIdWrapper {
-    fn deserialize(buf: &mut &[u8]) -> std::io::Result<Self> {
-        let value = u8::deserialize(buf)?;
-        Ok(SpecIdWrapper(unsafe { std::mem::transmute(value) }))
-    }
-
-    fn deserialize_reader<R: std::io::Read>(reader: &mut R) -> Result<Self, std::io::Error> {
-        let mut buf = vec![];
-        reader.take(1).read_to_end(&mut buf).unwrap(); // Read 1 bytes for a u8
-        let mut slice = buf.as_slice();
-        SpecIdWrapper::deserialize(&mut slice)
-    }
-}
 
 #[derive(Debug, Clone, PartialEq, BorshSerialize, BorshDeserialize)]
 pub struct EvmChainCfg {
